\name{gui}
\alias{exitPgm}
\alias{closePhytoImage}

\title{ The PhytoImage GUI (Graphical User Interface) }
\description{
  These function manage GUI parts, in addition to the GUI functions in the ZooImage package.
}
\usage{
exitPgm()
closePhytoImage()
}

\value{
  \code{exitPgm()} and \code{closePhytoImage()} unload the phytoimage/zooimage packages
  (for instance, to allow updating them).
}

\author{ Philippe Grosjean <Philippe.Grosjean@umons.ac.be> }

\keyword{ utilities }
