\name{phytoimage-package}
\alias{phytoimage-package}
\alias{phytoimage}
\docType{package}
\title{
PhytoImage: analysis of digital phytoplankton images.
}
\description{
PhytoImage is a free (open source) solution for analyzing digital
images of phytoplankton. In combination with ImageJ, a free image analysis
system, it processes digital images, measures individuals, trains for
automatic identification of taxa, and finally, measures phytoplankton samples
(abundances, total and partial size spectra or biomasses, etc.).

Most PhytoImage functions are included in the zooimage package. Please,
refer to it for the online help of these functions.
}
\details{
\tabular{ll}{
Package: \tab phytoimage\cr
Type: \tab Package\cr
Version: \tab 1.2-0\cr
Date: \tab 2007-05-24\cr
License: \tab GPL 2 or above at your convenience.\cr
}
For any series that you analyses with PhytoImage, the only constraints are:
- To send main metadata and a short description of your series, and the way it
  is analyzed to Philippe.Grosjean@umh.ac.be (this will be included to the
  PhytoImage web site (in order to keep track of its use).
- Everytime you publish results that use PhytoImage, you must place a reference
  to the PhytoImage web site (http://www.sciviews.org/phytoimage) in your publication.
  For papers, send also a reprint to Philippe.Grosjean@umh.ac.be, preferrably
  as a PDF file.
}
\author{
Philippe Grosjean & Kevin Denis, Numerical Ecology of Aquatic Systems,
Mons-Hainaut University, Belgium.

Maintainer: Philippe Grosjean <Philippe.Grosjean@umh.ac.be>
}
\references{
Not yet!
}
\keyword{ package }

\examples{
	# TODO...
}
