\name{phytoimage-package}
\alias{phytoimage-package}
\alias{phytoimage}
\docType{package}
\title{
PhytoImage: analysis of digital phytoplankton images.
}
\description{
PhytoImage is a free (open source) solution for analyzing digital
images of phytoplankton. In combination with ImageJ, a free image analysis
system, it processes digital images, measures individuals, trains for
automatic identification of taxa, and finally, measures phytoplankton samples
(abundances, total and partial size spectra or biomasses, etc.).

Most PhytoImage functions are included in the zooimage package. Please,
refer to it for the online help of these functions.
}
\details{
\tabular{ll}{
Package: \tab phytoimage\cr
Type: \tab Package\cr
Version: \tab 3.0-0\cr
Date: \tab 2012-05-07\cr
License: \tab GPL 2 or above at your convenience.\cr
}
Everytime you publish results that use PhytoImage, you must place a reference
to the Zoo/PhytoImage web site (http://www.sciviews.org/zooimage) in your publication.
For papers, send also a reprint to Philippe.Grosjean@umons.ac.be, preferrably
as a PDF file.
}
\author{
Philippe Grosjean & Kevin Denis, Numerical Ecology of Aquatic Systems,
Mons University, Belgium.

Maintainer: Philippe Grosjean <Philippe.Grosjean@umons.ac.be>
}
\references{
Not yet!
}
\keyword{ package }

\examples{
	# TODO...
}
